Add custom file server for serving static assets with correct MIME types
